Add files via upload
A wedding website should be romantic and personal, sharing the love story of the couple and all the details of their special day. It should include the wedding date, venue, schedule, and RSVP options, along with photos and a heartfelt message. Features like a guestbook and gift registry add a personal touch to make the experience memorable.